Adjust paths for EDITOR and PAGER environment variables in the fixit shell.  This would break them for floppy based fixit mode which still seems to use /stand except that vi(1) and more(1) don't exist in /stand on the fixit floppy so it really doesn't matter what these settings are there.  These paths work for CD-based fixit mode.
This is just "temporary" and on its way to 7.0-REL.  I'm too chicken to
make what is probably the correct adjustment this close to release.  It
seems /mnt2 is just a symbolic link, and stuff really gets mounted as
/dist.  Adjusting paths to that for 8.X is probably the right thing to
do and I'll try that after 7.0 is done.

Noticed by:	gallatin
MFC after:	1 day