MFC of r193498
[patch] Portuguese timedef

In Portuguese, the names of the days of the week are not capitalized.
Also there is always a dash before "feira" in the names of the days.
For example: "segunda-feira" and not "segunda feira" (which has a
completely different meaning).

x_fmt is not correct either. The date separator should not be a dot
but a slash. Example: 31/12/2005 if far more used in Portugal than
31.12.2005.

References:
- a Portuguese online dictionary http://priberam.pt/dlpo/dlpo.aspx
- http://answers.com/days_of_the_week (there are translations to
	various languages, including Portuguese, at the bottom of the
	page)
- http://en.wikipedia.org/wiki/Week-day_names (there are translations
	to various languages, including Portuguese, at the bottom of the
	page)
- a Portuguese style guide
	http://www.publico.clix.pt/nos/livro_estilo/16d-palavras.html
	("datas" section)

PR:           conf/58595
Submitted by: Chris Stenton <jacs@gnome.co.uk>