MFS 9->8 r217491 with slight modification:
Reduce the amount of text users need to read on each screen. Also try and
make it more relevant to modern systems: for example users will be
familiar with the MBR but possibly less so with MS-DOS.

After installation has finished don't return to the main menu but exit.

Slight modification:

Ignore the last sentence (need feedback from brucec before carrying that back
to 8 where sysinstall is still the main installer).