MFC 263779, except for the ATF test change.
Correct ARP update handling when the routes for network interfaces are
restricted to a single FIB in a multifib system.

Restricting an interface's routes to the FIB to which it is assigned (by
setting net.add_addr_allfibs=0) causes ARP updates to fail with "arpresolve:
can't allocate llinfo for x.x.x.x".  This is due to the ARP update code hard
coding it's lookup for existing routing entries to FIB 0.

sys/netinet/in.c:
        When dealing with RTM_ADD (add route) requests for an interface, use
        the interface's assigned FIB instead of the default (FIB 0).

sys/netinet/if_ether.c:
        In arpresolve(), enhance error message generated when an
        lla_lookup() fails so that the interface causing the error is
        visible in logs.

PR:             kern/167947