Revisit this old board with 64MB of RAM. Comment out usb entirely, since it isn't used for my application. Add back the md device since it's needed for NanoBSD support. Add in many of the small memory footprint options from the access points.
With these changes we go from having ~8MB to having ~20MB free,
though free + inactive only goes from ~35MB to ~42MB. We can
also boot a nanobsd image mostly (I had to hand tweak what was
built to represent the final goal).

Move the FDT stuff to the top. We're almost ready to pull the trigger
to moving over to FDT, but something in the MCI driver is freaking out
when we do and that needs fixing first.